fix(sd-storage): tighten guards, skip redundant power-down, throttle icon serving, widen usage to uint64
- sd_storage.h/.cpp: drop the no-op #else stub of sd_storage_mount() and
  wrap the declaration in #if USE_SD_STORAGE so the function only exists
  when SD storage is selected. All callers are already guarded.
- sd_probe.cpp: when USE_SD_STORAGE is also enabled, leave the SD card
  powered after a successful probe so sd_storage_mount() can reuse it
  immediately instead of cycling power 10 ms later.
- web_portal_icons.cpp: serve icon files via sendFileThrottled() instead
  of request->send(Storage, ...) to avoid WiFi TX buffer exhaustion on
  ESP-Hosted (same fix already applied to the FsIndexedStore JSON path).
- fs_health + storage: widen fs_health_set_storage_usage() and the
  FSHealthStats counters from uint32_t to uint64_t so SD cards larger
  than 4 GiB report correct total/used bytes. The HA discovery sensor
  already declares the unit as 'B', so the wider numeric value remains
  semantically correct end-to-end.

+++ src/app/fs_health.cpp
@@ -34,7 +34,7 @@ void fs_health_init() {
 		detect_partitions();
 }
 
-void fs_health_set_storage_usage(uint32_t used_bytes, uint32_t total_bytes) {
+void fs_health_set_storage_usage(uint64_t used_bytes, uint64_t total_bytes) {
 		// Treat this as a one-way latch: once mounted, keep reporting mounted.
 		g_stats.storage_mounted = true;
 		g_stats.storage_used_bytes = used_bytes;
